welcomes: fixed lethal bugs

+++ tg_bot/modules/welcome.py
@@ -182,8 +182,7 @@ def new_member(bot: Bot, update: Update):
 
                 #Safe mode
                 newMember = chat.get_member(int(new_mem.id))
-                c = False
-                if welc_mutes == "on" and (newMember.can_send_messages is None or newMember.can_send_messages):
+                if welc_mutes == "on" and ((newMember.can_send_messages is None or newMember.can_send_messages)):
                     buttonMsg = msg.reply_text("Click the button below to prove you're human, otherwise you will be kicked in about " + str(time_value) + " seconds. Thank you!",
                          reply_markup=InlineKeyboardMarkup([[InlineKeyboardButton(text="I'm not a bot!", 
                          callback_data="userverify_({})".format(new_mem.id))]]))
@@ -195,14 +194,13 @@ def new_member(bot: Bot, update: Update):
                     time.sleep(10)
                     newMember = chat.get_member(int(new_mem.id)) #I MISS POINTERS
                     i = 10
-                    while i < time_value and newMember.can_send_messages == False:
+                    while i < time_value and newMember.can_send_messages == False and (newMember.status is not "left"):
                         i+=1
                         time.sleep(1)
-                    if newMember.can_send_messages == False:
-                        chat.kick_member(new_mem.id)
+                    if newMember.can_send_messages == False and newMember.status is not "left":
+                        bantime = int(time.time()) + 35
+                        chat.kick_member(new_mem.id, until_date=bantime)
                         buttonMsg.delete()
-                        time.sleep(10)
-                        chat.unban_member(new_mem.id)
                         
             delete_join(bot, update)
